+++ src/components/Checklist/Checklist.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Item } from '../Item/Item';
 
-export const Checklist = ({ cards, onClickDetail, isShowDetailItem, titleValue }) => {
+export const Checklist = ({ cards, titleValue }) => {
 
   return (
     <ul className="relative overflow-visible">
@@ -13,8 +13,6 @@ export const Checklist = ({ cards, onClickDetail, isShowDetailItem, titleValue }
             src={oneCard.image}
             labels={oneCard.labels}
             key={oneCard.id}
-            onClickDetail={onClickDetail}
-            isShowDetailItem={isShowDetailItem}
           />
         ))
       }
