+++ src/index.jsx
@@ -2,10 +2,13 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App.jsx';
 import './index.css';
-import { LabelProvider } from './context/LabelContext.jsx';
+import { DetailProvider } from './context/DetailContext';
+import { LabelProvider } from './context/LabelContext';
 
 ReactDOM.createRoot(document.getElementById('root')).render(
-  <LabelProvider>
-    <App />
-  </LabelProvider>
+  <DetailProvider>
+    <LabelProvider>
+      <App />
+    </LabelProvider>
+  </DetailProvider>
 )
